observes\fix(back): #5993.16 big ints
- use big ints by default for avoiding NumericValueOutOfRange

+++ observes/singer/target_redshift/target_redshift/data_types.py
@@ -28,7 +28,7 @@ class NullOrEmptyType(Exception):
 
 raw_type_map: Dict[str, RedshiftDataType] = {
     "boolean": RedshiftDataType.BOOLEAN,
-    "integer": RedshiftDataType.INTEGER,
+    "integer": RedshiftDataType.BIGINT,
     "number": RedshiftDataType.DOUBLE_PRECISION,
     "string": RedshiftDataType.VARCHAR,
 }
